+++ pkg/utils/timestamp.go
@@ -5,23 +5,26 @@ import (
 )
 
 // NormalizeFloatTimestamp normalizes a floating-point timestamp to seconds and nanoseconds.
-func NormalizeFloatTimestamp(timestamp float64) (int64, int64) {
+func NormalizeFloatTimestamp(timestamp float64) (int64, int32) {
 	switch {
 	case timestamp >= 1_000_000_000_000_000_000:
 		// timestamp in nanoseconds
 		sec := int64(timestamp / 1_000_000_000)
 		nsec := int64(timestamp) % 1_000_000_000
-		return sec, nsec
+		//nolint: gosec // ignore int64 to int32 conversion
+		return sec, int32(nsec)
 	case timestamp >= 1_000_000_000_000_000:
 		// timestamp in microseconds
 		sec := int64(timestamp / 1_000_000)
 		nsec := int64(timestamp) % 1_000_000 * 1_000
-		return sec, nsec
+		//nolint: gosec // ignore int64 to int32 conversion
+		return sec, int32(nsec)
 	case timestamp >= 1_000_000_000_000:
 		// timestamp in milliseconds
 		sec := int64(timestamp / 1_000)
 		nsec := int64(timestamp) % 1_000 * 1_000_000
-		return sec, nsec
+		//nolint: gosec // ignore int64 to int32 conversion
+		return sec, int32(nsec)
 	default:
 		// timestamp in seconds
 		return int64(timestamp), 0
@@ -31,7 +34,7 @@ func NormalizeFloatTimestamp(timestamp float64) (int64, int64) {
 // TimeFromFloat converts a floating-point timestamp to a time.Time instance.
 func TimeFromFloat(timestamp float64) time.Time {
 	sec, nsec := NormalizeFloatTimestamp(timestamp)
-	return time.Unix(sec, nsec)
+	return time.Unix(sec, int64(nsec))
 }
 
 // FloatSecFromTime converts a time.Time instance to a floating-point timestamp in seconds.
